# docs/contracts/openapi/platform-core.yaml
openapi: 3.0.3
info:
  title: 1Trade Platform-Core API (auth, accounts, RBAC, keys, billing)
  version: 1.6.0
  description: >
    Identity for the whole platform. Owned by `platform-core`. Every other service authenticates
    inbound requests by verifying the JWT this service issues (see JwtClaims) — that's how each
    service resolves tenant_id / sub_account_id / roles / is_paper. SAML/SCIM/2FA are M4 additions
    (this v1 covers email+password, OAuth, API keys, and the account/org surface). See
    docs/plans/features/F02-auth-and-sso.md and F03-accounts-orgs-rbac.md.

    **v1.6.0 (2026-09-28, team management).** Members (list; admins remove — never themselves, never
    the last admin), invitations (admins invite an email with roles and optionally a sub-account; the
    emailed token is single use, 7 days, and accepting it creates a verified user in the inviting
    tenant and signs them in; one account per email), and sub-accounts (teams with their own ledger
    balances: members placed in one carry `sub_account_id` in their JWT and are billed to it; admin
    or billing funds or drains it from the main balance through credit.yaml v1.4 transfer). Every
    change is audited. `/v1/auth/me` returns `sub_account_id`.
servers:
  - { url: "https://api.1trade.io", description: prod }
  - { url: "http://localhost:8080", description: "local (k3d ingress)" }
...
paths:
  /v1/auth/signup:
    post:
      tags: [auth]
      summary: Self-serve signup — creates an individual tenant + admin user, returns a JWT (auto-login).
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
                tenant_name: { type: string }
      responses:
        '201':
          description: Created + authenticated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  expires_at: { type: string, format: date-time }
        '409': { description: Email already registered., content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Email + password login. Returns a JWT (see JwtClaims).
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Authenticated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string, description: "JWT bearer token" }
                  expires_at: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/auth/oauth/{provider}:
    get:
      tags: [auth]
      summary: Start the OAuth browser flow (google | github | microsoft). Redirects to the IdP.
      operationId: oauthStart
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string, enum: [google, github, microsoft] } }
        - { name: redirect_uri, in: query, schema: { type: string } }
      responses:
        '302': { description: Redirect to the identity provider. }
  /v1/auth/oauth/{provider}/callback:
    get:
      tags: [auth]
      summary: OAuth callback. Exchanges the code, then redirects back with a session set.
      operationId: oauthCallback
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string } }
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '302': { description: Authenticated; redirect to the app. }
  /v1/auth/logout:
    post:
      tags: [auth]
      summary: Invalidate the current session.
      operationId: logout
      security: [{ tenantJwt: [] }]
      responses:
        '204': { description: Logged out. }
  /v1/auth/me:
    get:
      tags: [auth]
      summary: The current identity (decoded from the bearer token).
      operationId: me
      security: [{ tenantJwt: [] }]
      responses:
        '200':
          description: Current identity.
          content: { application/json: { schema: { $ref: '#/components/schemas/Identity' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/auth/keys:
    get:
      tags: [auth]
      summary: List the tenant's API keys (metadata only; never the secret).
      operationId: listKeys
      security: [{ tenantJwt: [] }]
      responses:
        '200':
          description: Keys.
          content:
            application/json:
              schema: { type: object, properties: { keys: { type: array, items: { $ref: '#/components/schemas/ApiKey' } } } }
    post:
      tags: [auth]
      summary: Create a scoped API key. The secret is returned ONCE.
      operationId: createKey
      security: [{ tenantJwt: [] }]
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name: { type: string }
                scopes: { type: array, items: { type: string }, description: "e.g. inference:read, compute:write" }
      responses:
        '201':
          description: Created; `secret` shown once.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiKey'
                  - type: object
                    properties: { secret: { type: string, description: "shown once — store it now" } }
  /v1/auth/keys/{id}:
    delete:
      tags: [auth]
      summary: Revoke an API key.
      operationId: revokeKey
      security: [{ tenantJwt: [] }]
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '204': { description: Revoked. }
  /v1/auth/keys/introspect:
    post:
      tags: [auth]
      summary: >
        Internal — resolve a raw API key to its tenant/scopes/is_paper. Called service-to-service
        (e.g. the inference gateway) to authenticate a customer's `exk_...` key. Not for end users.
      operationId: introspectKey
      security: [{ serviceToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [api_key]
              properties:
                api_key: { type: string, description: "the raw key, e.g. exk_..." }
      responses:
        '200':
          description: Resolved principal.
          content: { application/json: { schema: { $ref: '#/components/schemas/KeyPrincipal' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Unknown or revoked key.
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
  /v1/account/tenants/{id}:
    get:
      tags: [accounts]
      summary: Tenant details.
      operationId: getTenant
      security: [{ tenantJwt: [] }]
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '200': { description: Tenant., content: { application/json: { schema: { $ref: '#/components/schemas/Tenant' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/account/orgs/{id}/users:
    get:
      tags: [accounts]
      summary: List users in an org (with roles).
      operationId: listOrgUsers
      security: [{ tenantJwt: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: Users.
          content:
            application/json:
              schema:
                type: object
                properties:
                  users: { type: array, items: { $ref: '#/components/schemas/User' } }
                  next_cursor: { type: string, nullable: true }
  /v1/auth/verify:
    post:
      tags: [auth]
      summary: Verify an email address with the one-time token (the token is the credential; no bearer).
      operationId: verifyEmail
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [token], properties: { token: { type: string } } } } }
      responses:
        '200': { description: Verified., content: { application/json: { schema: { type: object, properties: { verified: { type: boolean } } } } } }
        '400': { description: Invalid or used token., content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  /v1/auth/verify/resend:
    post:
      tags: [auth]
      summary: >-
        Resend an email-verification link. Authenticated → for the caller (dev returns the token).
        Unauthenticated with {email} → for that address if it exists and is unverified; the response is
        a constant {sent:true} that never reveals whether the email is registered (used by the login
        screen's "resend verification" when sign-in is blocked by the verification gate).
      operationId: resendVerify
      security: [{ tenantJwt: [] }, {}]   # authed (caller) OR anonymous (by email)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email, description: "anonymous mode — resend for this address" }
      responses:
        '200': { description: Sent — constant response (no account enumeration). }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/billing/budget:
    get:
      tags: [billing]
      summary: The tenant's monthly credit budget (null if none set).
      operationId: getBudget
      security: [{ tenantJwt: [] }]
      responses:
        '200': { description: Budget. }
    put:
      tags: [billing]
      summary: Set the tenant's monthly budget (billing/admin). Drives 50/80/100% alerts. Audited.
      operationId: setBudget
      security: [{ tenantJwt: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [credit_type, monthly_limit], properties: { credit_type: { type: string }, monthly_limit: { type: string, description: "fixed-point decimal string" } } } } }
      responses:
        '200': { description: Updated. }
        '403': { description: Insufficient role., content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
...
  /v1/account/orgs:
    post:
      tags: [accounts]
      summary: Create an org under the tenant (admin-only). Audited.
      operationId: createOrg
      security: [{ tenantJwt: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string } } } } }
      responses:
        '201': { description: Created., content: { application/json: { schema: { $ref: '#/components/schemas/Org' } } } }
        '403': { description: Insufficient role., content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    get:
      tags: [accounts]
      summary: List the tenant's orgs.
      operationId: listOrgs
      security: [{ tenantJwt: [] }]
      responses:
        '200':
          description: Orgs.
          content: { application/json: { schema: { type: object, properties: { orgs: { type: array, items: { $ref: '#/components/schemas/Org' } } } } } }
  /v1/account/users/{id}/roles:
    put:
      tags: [accounts]
      summary: Assign a user's roles within the tenant (admin-only). Audited (before/after).
      operationId: assignRoles
      security: [{ tenantJwt: [] }]
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [roles], properties: { roles: { type: array, items: { $ref: '#/components/schemas/Role' } } } } } }
      responses:
        '200': { description: Updated. }
        '403': { description: Insufficient role., content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { description: Unknown role., content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  /v1/account/audit:
    get:
      tags: [accounts]
      summary: The tenant's audit log, newest first (admin-only). Every sensitive action is recorded.
      operationId: listAudit
      security: [{ tenantJwt: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 50 } }
      responses:
        '200':
          description: Audit entries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        id:          { type: string, format: uuid }
                        actor_id:    { type: string, format: uuid, nullable: true }
                        action:      { type: string }
                        target_type: { type: string, nullable: true }
                        target_id:   { type: string, nullable: true }
                        before:      { type: object, additionalProperties: true, nullable: true }
                        after:       { type: object, additionalProperties: true, nullable: true }
                        is_paper:    { type: boolean }
                        created_at:  { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Insufficient role., content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  /v1/account/kyc:
    get:
      tags: [accounts]
      summary: The caller's tenant KYC status (gates real-money purchases).
      operationId: getKyc
      security: [{ tenantJwt: [] }]
      responses:
        '200':
          description: KYC status.
          content: { application/json: { schema: { $ref: '#/components/schemas/Kyc' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [accounts]
      summary: >
        Submit identity verification (KYC/AML) for the caller's tenant. Captures minimal PII (legal
        name · ISO-3166 country · entity type). In prod the submission lands `pending` for review; in
        dev/sandbox it is auto-verified. Allowed only from unverified/rejected.
      operationId: submitKyc
      security: [{ tenantJwt: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [legal_name, country, entity_type]
              properties:
                legal_name:  { type: string }
                country:     { type: string, description: "ISO 3166-1 alpha-2" }
                entity_type: { type: string, enum: [individual, business] }
      responses:
        '200':
          description: The resulting KYC state (verified in dev, else pending).
          content: { application/json: { schema: { $ref: '#/components/schemas/Kyc' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: "Already pending or verified (code kyc_not_submittable).", content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '422': { description: Invalid submission., content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  /v1/account/kyc/{tenant_id}/decision:
    post:
      tags: [accounts]
      summary: >
        Apply a compliance decision (verified|rejected) to a pending tenant submission. INTERNAL —
        authorized by the service-to-service token, never an end-user bearer (a tenant cannot verify
        itself). The prod manual-review / IDV-vendor callback path.
      operationId: reviewKyc
      security: [{ serviceToken: [] }]
      parameters:
        - { name: tenant_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [verified, rejected] }
      responses:
        '200':
          description: Decision applied.
          content: { application/json: { schema: { type: object, properties: { tenant_id: { type: string, format: uuid }, status: { $ref: '#/components/schemas/KycStatus' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: "No pending submission for this tenant.", content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
...
  /v1/billing/purchases:
    get:
      tags: [billing]
      summary: The tenant's purchase history.
      operationId: listPurchases
      security: [{ tenantJwt: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200':
          description: Purchases.
          content:
            application/json:
              schema:
                type: object
                properties:
                  purchases: { type: array, items: { $ref: '#/components/schemas/Purchase' } }
components:
  securitySchemes:
    tenantJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        The platform JWT. Every service verifies it (signature + exp) and reads JwtClaims to
        authorize by tenant_id / sub_account_id / roles, and to enforce is_paper isolation.
    apiKey:
      type: http
      scheme: bearer
      description: "A scoped API key (for programmatic clients, e.g. the inference API)."
    serviceToken:
      type: http
      scheme: bearer
      description: "Service-to-service token for internal endpoints (e.g. the inference gateway introspecting an API key)."
...
security:
  - tenantJwt: []